+++ platformio.ini
@@ -16,7 +16,7 @@ default_envs = CUSTOMBOARD
 platform = espressif32
 framework = arduino
 version = 0.1.8_dev
-revision = 46
+revision = 47
 monitor_speed = 115200
 monitor_rts = 0
 monitor_dtr = 0
@@ -40,6 +40,7 @@ lib_deps =
 	lvgl/lvgl@^8.3.9
 	lovyan03/LovyanGFX@^1.1.9
 	jchristensen/Timezone@^1.2.4
+	bblanchon/StreamUtils@^1.7.3
 
 [esp32_common]
 platform = ${common.platform}
@@ -65,7 +66,6 @@ lib_deps =
 	adafruit/Adafruit HMC5883 Unified@^1.2.1
 	adafruit/Adafruit BusIO@^1.14.5
 	adafruit/Adafruit BME280 Library@^2.2.2
-	bblanchon/StreamUtils@^1.7.3
 build_flags = 
 	${common.build_flags}
 	-DBOARD_HAS_PSRAM
@@ -84,7 +84,6 @@ board_build.f_cpu = 240000000L
 lib_deps = 
 	${common.lib_deps}
 	bolderflight/Bolder Flight Systems MPU9250@^1.0.2
-	bblanchon/StreamUtils@^1.7.3
 build_flags = 
 	${common.build_flags}
 	-DBOARD_HAS_PSRAM
